pyproject.toml
[tool.poetry]
name = "matlab-parser"
version = "0.1.0"
description = ""
authors = ["Samuel Grayson <sam@samgrayson.me>"]
license = "MIT"
packages = [
    { include = "matlab-parser" },
]

[tool.poetry.scripts]
matlab-parser = 'matlab-parser.main:main'

[tool.poetry.dependencies]
python = "^3.7"
antlr4-python3-runtime = "^4.9.2"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
ipython = "^7.27.0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
